(PC-17817) chore(search): move atoms to components folder
