+++ src/features/search/components/Buttons/FilterButton/SingleFilterButton.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { SingleFilterButton } from 'features/search/atoms/Buttons/FilterButton/SingleFilterButton'
+import { SingleFilterButton } from 'features/search/components/Buttons/FilterButton/SingleFilterButton'
 import { fireEvent, render } from 'tests/utils'
 
 describe('SingleFilterButton', () => {
