+++ src/features/search/components/Buttons/HiddenNavigateToSuggestionsButton.tsx
