+++ src/features/search/components/CenteredSection.tsx
