+++ src/features/search/components/SearchBox.tsx
@@ -16,8 +16,8 @@ import { v4 as uuidv4 } from 'uuid'
 import { useAppSettings } from 'features/auth/settings'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
-import { FilterButton } from 'features/search/atoms/Buttons/FilterButton/FilterButton'
-import { HiddenNavigateToSuggestionsButton } from 'features/search/atoms/Buttons/HiddenNavigateToSuggestionsButton'
+import { FilterButton } from 'features/search/components/Buttons/FilterButton/FilterButton'
+import { HiddenNavigateToSuggestionsButton } from 'features/search/components/Buttons/HiddenNavigateToSuggestionsButton'
 import { useAppliedFilters } from 'features/search/components/useAppliedFilters'
 import { useHasPosition } from 'features/search/components/useHasPosition'
 import { LocationModal } from 'features/search/pages/LocationModal'
