+++ src/features/search/components/SearchResults.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components/native'
 import { SearchGroupNameEnumv2 } from 'api/gen'
 import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { Hit, NoSearchResult, NumberOfResults } from 'features/search/atoms'
-import { SingleFilterButton } from 'features/search/atoms/Buttons/FilterButton/SingleFilterButton'
+import { SingleFilterButton } from 'features/search/components/Buttons/FilterButton/SingleFilterButton'
 import { AutoScrollSwitch } from 'features/search/components/AutoScrollSwitch'
 import { ScrollToTopButton } from 'features/search/components/ScrollToTopButton'
 import { FILTER_TYPES, useAppliedFilters } from 'features/search/components/useAppliedFilters'
